Init rewrite of bulkMenu.js
